fix btc only bug
