+++ src/ui/lv_i18n/lv_i18n_api.c
@@ -14,11 +14,9 @@ const LanguageInfo_t g_languageList[] = {
     {LANG_RU, "ru"},
     {LANG_ZH_CN, "zh-CN"},
     {LANG_KO, "ko"},
-#ifndef BTC_ONLY
     {LANG_ES, "es"},
     {LANG_DE, "de"},
     {LANG_JA, "ja"},
-#endif
 };
 
 static uint8_t g_curLangIndex = LANG_EN;
